requirements.txt
flake8
pylint
coverage
pyspark==3.5.0
pandas==2.1.0
pyarrow==14.0.1
numpy==1.24.3
python-dotenv==1.0.0
sqlalchemy==1.4.46
psycopg2-binary==2.9.6
pytest==7.3.1
